bodypix_tflite/requirements.txt
Pillow==7.2.0
opencv-python==4.5.4.60
numpy==1.20.1


